requirements.txt
fastapi
uvicorn[standard]
numpy
faiss-cpu
transformers
sentence-transformers
torch
requests
python-multipart
pydantic
faster-whisper
TTS
